3D Models, Sound, Actors don't go through walls
I downloaded 3D Spy vs Spy models, and also some random secret agent
model, but ended up just using the spy vs spy models. Added music that
shuffles between tracks. Fixed a bug that allowed actors to go through
walls, and added checks so you can't make a waypoint on the other side
of a wall. At some point I plan to implement A* pathfinding using nodes
at the corners of walls, but that's not going to be for a while. I also
moved some folders around, which is why so many files got changed in
this commit.